Convert variables UI to TS
